+++ views/email/components/commonCss.ejs
@@ -107,10 +107,7 @@
 	footer {
 		padding: 1em;
 		font-size: 0.75rem;
-		display: flex;
-		flex-direction: column;
-		align-items: center;
-		justify-content: center;
+		text-align: center;
 	}
 
 	footer > * {
@@ -122,10 +119,17 @@
 		display: flex;
 		align-items: center;
 		justify-content: center;
+		flex-wrap: wrap;
+		margin: 0 auto;
+		text-align: center;
+	}
+
+	footer ul li {
+		margin: 0 auto;
 	}
 
 	footer nav {
-		width: clamp(200px, 100%, 320px);
+		width: clamp(200px, 100%, 400px);
 	}
 
 	footer nav li {
