+++ views/email/components/footer.ejs
@@ -1,2 +1,50 @@
-    </body>
+<footer class="text-center">
+    <p class="address">
+        © Sundar Clinic, 1195A, Nehru Street, Bangalore-Chennai Highway,
+        Pappanchatiram, Chennai - 600123.
+    </p>
+    <center style="margin-left: auto; margin-right: auto">
+        <nav>
+            <ul>
+                <li>
+                    <a href="<%=reactAppUrl%>/termsOfService"
+                        >Terms of Service</a
+                    >
+                </li>
+                <li>
+                    <a href="<%=reactAppUrl%>/privacyPolicy"
+                        >Privacy Policy</a
+                    >
+                </li>
+                <li><a href="<%=reactAppUrl%>/contact">Contact</a></li>
+            </ul>
+            <ul class="socials">
+                <li>
+                    <a href="https://www.instagram.com/sundarclinic/"
+                        ><img
+                            src="https://sundar-clinic.herokuapp.com/icons/instagram.png"
+                            alt="Instagram Icon"
+                    /></a>
+                </li>
+                <li>
+                    <a
+                        href="https://api.whatsapp.com/send?phone=8939881708&text=Hello%20there!%20"
+                        ><img
+                            src="https://sundar-clinic.herokuapp.com/icons/whatsapp.png"
+                            alt="Whatsapp Icon"
+                    /></a>
+                </li>
+                <li>
+                    <a href="mailto:sundarclinic@gmail.com"
+                        ><img
+                            src="https://sundar-clinic.herokuapp.com/icons/gmail.png"
+                            alt="Gmail Icon"
+                    /></a>
+                </li>
+            </ul>
+        </nav>
+    </center>
+</footer>
+</div>
+   </body>
 </html>
