+++ views/email/welcomeAndVerify.ejs
@@ -21,55 +21,13 @@
 			>Verify Account</a
 		>
 		<p>Alternatively click or copy and paste this link in your browser:</p>
-		<a href=""><%=reactAppUrl%>/verifyAccount/authToken=<%=token%></a>
+		<a href="<%=reactAppUrl%>/verifyAccount/authToken=<%=token%>"
+			><%=reactAppUrl%>/verifyAccount/authToken=<%=token%></a
+		>
 		<hr />
 		<p class="text-center">
 			Your healthcare partners at Sundar Clinic.
 			<span class="text-red">❤</span>
 		</p>
 	</div>
-	<footer class="text-center">
-		<p class="address">
-			© Sundar Clinic, 1195A, Nehru Street, Bangalore-Chennai Highway,
-			Pappanchatiram, Chennai - 600123.
-		</p>
-		<nav>
-			<ul>
-				<li>
-					<a href="<%=reactAppUrl%>/termsOfService"
-						>Terms of Service</a
-					>
-				</li>
-				<li>
-					<a href="<%=reactAppUrl%>/privacyPolicy">Privacy Policy</a>
-				</li>
-				<li><a href="<%=reactAppUrl%>/contact">Contact</a></li>
-			</ul>
-		</nav>
-		<ul class="socials">
-			<li>
-				<a href="https://www.instagram.com/sundarclinic/"
-					><img
-						src="https://sundar-clinic.herokuapp.com/icons/instagram.png"
-						alt="Instagram Icon"
-				/></a>
-			</li>
-			<li>
-				<a
-					href="https://api.whatsapp.com/send?phone=8939881708&text=Hello%20there!%20"
-					><img
-						src="https://sundar-clinic.herokuapp.com/icons/whatsapp.png"
-						alt="Whatsapp Icon"
-				/></a>
-			</li>
-			<li>
-				<a href="mailto:sundarclinic@gmail.com"
-					><img
-						src="https://sundar-clinic.herokuapp.com/icons/gmail.png"
-						alt="Gmail Icon"
-				/></a>
-			</li>
-		</ul>
-	</footer>
-</div>
 <%- include ("./components/footer.ejs") %>
